Add support for OpenImageIO default_thread_pool_shutdown() (#35)
Adding support for calling default_thread_pool_shutdown() on
Windows. This should fix the hangs during shutdown that are
happening on that platform. This function is expected to be
included in the 2.5.3 release of OpenImageIO. This function
has also been backported to the Natron OpenImageIO MinGW
package on Windows and signalled via a
NATRON_OIIO_HAS_DEFAULT_THREAD_POOL_SHUTDOWN define.